Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultanious Release. Add a profile that is active by
default so that default behaviour is unchanged however allows the
autorelease project to disable building this module.

Change-Id: Icf1f8c9e05c5667e5e7483783f2651ad89fc0eac
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>