Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultanious Release. Add a profile that is active by
default so that default behaviour is unchanged however allows the
autorelease project to disable building this module.

Issue: RELENG-86
Change-Id: I822a1f08d25362056e9da401e05dc88b2a2cd20b
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>